Remove weight and style mixins. (#209)
* Remove weight and style mixins.

Removes the following mixins in favour of updating the arguments of related
mixins [see the v6 proposal](https://github.com/Financial-Times/o-typography/issues/203):
- oTypographySansBold
- oTypographyDisplayBold
- oTypographySerifBold
- oTypographySerifItalic
- oTypographyBold
- oTypographyItalic

Instead `oTypographySans`, `oTypographyDisplay`, and `oTypographySerif` accept a
`$weight` and `$style` argument. If the given weight or style does not have
a corresponding font in Origami an error is thrown. The `$progressive` parameter
has been moved within an `$opts` map to optionally exclude the sizes of a
fallback font, and in addition a font-family option has been added to exclude
the font family; so these mixins may be used to modify a font size without
repeating the font-family declaration. See the tests and SassDoc examples.

README.md is out of date and MIGRATION.md has only a list of changes.
They will be updated with in a later commit.

* Update progressive loading classes to match our naming convention.

The class `o-typography--loading-sansBold` is now `o-typography--loading-sans-bold` and
`o-typography--loading-displayBold` is now `o-typography--loading-display-bold`.