Remove oTypographySize.
We introduced multiple font scales for specialist titles earlier
in the year, so the result of `oTypographySize` depends on the font.

Instead of passing a font type to `oTypographySize` use
`oTypographySans`, `oTypographySerif`, or `oTypographyDisplay` with
arguments instead. This also includes progressive font loading
styles for the updated size, where `oTypographySize` did not.