Bring master brand product headings closer to the current style.
As part of the major release we have made product headings the
default heading. We did not have time to review their visual
style thoroughly but removed excess heading levels 7 and 8.
This commit brings the style of the headings closer to the master
brands current style as apposed to the current internal brand
style.