Add `$o-typography-error-for-missing-fonts`.
Adds the option to error if a font variant (family/weight/style)
is used when a font face for that variant has not been included in
the project. By default no error is thrown unless the Sass variable
`$o-typography-error-for-missing-fonts` is set to a string (a
message to include in the error).

This will allow projects like PageKit to remove fonts and inform
dependent projects what to do with a custom error message.

We may want to make an error the default behaviour in a future major
version of o-typography.

Relates to: https://github.com/Financial-Times/o-typography/issues/249